refactor: stateless MCP 서버 1차 도구 구현
