+++ src/hwpx_mcp_server/core/__init__.py
@@ -1,0 +1,1 @@
+"""Stateless HWPX 코어 연산."""
