+++ src/hwpx_mcp_server/core/formatting.py
@@ -1,0 +1,1 @@
+"""Phase 1에서는 서식 전용 API를 노출하지 않습니다."""
