+++ tests/conftest.py
@@ -1,0 +1,4 @@
+from xml.etree import ElementTree as _ET
+import builtins
+
+builtins.ET = _ET
